+++ test/cypress/tests/modules/BlockEvents/Slash.cy.ts
@@ -19,15 +19,13 @@ describe('Slash keydown', function () {
         .click()
         .type('/');
 
-      cy.get('[data-cy="toolbox"]')
-        .get('.ce-popover')
+      cy.get('[data-cy="toolbox"] .ce-popover')
         .should('be.visible');
     });
 
     [
-      '{shift}',
-      '{alt}',
-      '{option}',
+      'ctrl',
+      'cmd',
     ].forEach((key) => {
       it(`should not open Toolbox if Slash pressed with ${key}`, () => {
         cy.createEditor({
@@ -46,10 +44,9 @@ describe('Slash keydown', function () {
         cy.get('[data-cy=editorjs]')
           .find('.ce-paragraph')
           .click()
-          .type(`${key}/`);
+          .type(`{${key}}/`);
 
-        cy.get('[data-cy="toolbox"]')
-          .get('.ce-popover')
+        cy.get('[data-cy="toolbox"] .ce-popover')
           .should('not.be.visible');
       });
     });
@@ -75,8 +72,7 @@ describe('Slash keydown', function () {
         .click()
         .type('/');
 
-      cy.get('[data-cy="toolbox"]')
-        .get('.ce-popover')
+      cy.get('[data-cy="toolbox"] .ce-popover')
         .should('not.be.visible');
 
       /**
@@ -110,8 +106,7 @@ describe('CMD+Slash keydown', function () {
       .click()
       .type('{cmd}/');
 
-    cy.get('[data-cy="block-tunes"]')
-      .get('.ce-popover')
+    cy.get('[data-cy="block-tunes"] .ce-popover')
       .should('be.visible');
   });
 });